Fix each_packet and each_data on Windows
On Windows, rb_thread_wait_fd and pcap_fileno don't play nice together.
What's worse this problem is often masked in Ruby 1.9 because
rb_thread_wait_fd is a noop if only one thread is running.

This fixes the each_packet and each_data methods on Windows when running
inside a thread (such as within a Timeout block).

On Windows Ruby >=1.9 pcap_getevent and WaitForSingleObject are used.

On Windows Ruby < 1.9 rb_thread_polling is used.
rb_thread_polling may cause some packets to be dropped, but this is
better than not capturing at all.